ml/requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
scikit-learn==1.3.2

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Computer Vision
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0

# NLP
sentencepiece==0.1.99
tokenizers>=0.14,<0.15
nltk==3.8.1

# Model Training
tensorboard==2.15.1
wandb==0.16.1

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
joblib==1.3.2

# Jupyter
jupyterlab==4.0.9
ipywidgets==8.1.1

# Model Export
onnx==1.15.0
onnxruntime==1.16.3

# Hyperparameter Tuning
optuna==3.5.0

# Data Augmentation
albumentations==1.3.1
imgaug==0.4.0